Componentized the Assigned.jsx of Admin - to separate the dialog into component
