+++ frontend/src/Pages/UserPages/Assigned.jsx
@@ -40,36 +40,6 @@ const Assigned = () => {
     fetchFeatures();
   }, [loggedUser._id]);
 
-  // const fetchFeatures = async () => {
-  //   try {
-  //     const response = await axios.get(
-  //       `/api/v1/features/assigned/${loggedUser._id}`
-  //     );
-  //     const featuresWithParentProject = await Promise.all(
-  //       response.data.features.map(async (feature) => {
-  //         const featureWithParentProject = { ...feature };
-  //         const parentProject = await axios.get(
-  //           `/api/v1/projects/${feature.parentProject}`
-  //         );
-  //         featureWithParentProject.parentProjectName =
-  //           parentProject.data.project.name;
-  //         featureWithParentProject.parentProjectDescription =
-  //           parentProject.data.project.description;
-  //         featureWithParentProject.parentProjectPriority =
-  //           parentProject.data.project.priority;
-  //         featureWithParentProject.parentProjectStatus = parentProject.data
-  //           .project.isDone
-  //           ? "Done"
-  //           : "Ongoing";
-  //         return featureWithParentProject;
-  //       })
-  //     );
-  //     setFeatures(featuresWithParentProject);
-  //   } catch (error) {
-  //     console.error("Error fetching features:", error);
-  //   }
-  // };
-
   const fetchFeatures = async () => {
     try {
       const response = await axios.get(
