--- frontend/src/data.js
@@ -1,14 +1,0 @@
-export const users = [
-  {
-    username: "user",
-    password: "password",
-    role: "user",
-    name: "Vlahdimyr Bambalan",
-  },
-  {
-    username: "admin",
-    password: "password",
-    role: "admin",
-    name: "Sharlize Love",
-  },
-];
